examples/park-ranger/pyproject.toml
[project]
name = "nearest-wildlife"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.14",
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.13.4",
    "click>=8.1.8",
    "feedparser>=6.0.11",
    "langchain-core>=0.3.74",
    "lxml>=6.0.0",
    "lxml-html-clean>=0.4.2",
    "mcp>=1.11.0",
    "newspaper3k>=0.2.8",
    "openai>=1.74.0",
    "openpipe-art[skypilot]",
    "python-dotenv>=1.1.1",
    "tenacity>=9.1.2",
    "weave>=0.51.56",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["just_the_facts*"]

[tool.uv.sources]
openpipe-art = { path = "../../", editable = true }

[dependency-groups]
dev = [
    "polars>=1.31.0",
    "ipywidgets>=8.1.6",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
]
